js/xpconnect/tests/mochitest/test_bug785096.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=390488
-->
<head>
  <title>Test for Bug 785096</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="test()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=785096">Mozilla Bug 785096</a>

<script class="testbody" type="text/javascript">

/** Test for Bug 785096 **/

SimpleTest.waitForExplicitFinish();

function test()
{
  var sel = document.createElementNS('http://www.w3.org/1999/xhtml', 'select');
  var option=document.createElement("option");
  option.text="Fubar";
  sel.options.add(option);
  try {
    SpecialPowers.Components.lookupMethod(sel.options, "add")(option);
    ok(true, "function call should not throw")
  } catch(e) {
    ok(false, "this call should just work without any exceptions");
  }
  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
